requirements.txt
pillow==10.3.0
moviepy==1.0.3